worked on storage.js

+++ lab-aaron/route/route-toy.js
@@ -67,15 +67,16 @@ module.exports = function(router) {
   router.delete('/api/toy', (req, res) => {
     debug('api/toy DELETE');
     if(req.url.query._id) {
-      storage.delete('toy', req.url.query._id)
-        .then(toy => {
-          res.writeHead(201, {'Content-Type': 'application/json'});
-          res.end();
-        })
-        .catch(e => {
-          console.error(e);
-          res.writeHead(400, `bad request: ${e.message}`);
-        });
+      if(!req.url.query._id){
+        res.writeHead(404);
+        res.write('bad request; request improperly forammted for delete');
+        res.end();
+        return;
+      }
+      storage.delete('toy', req.url.query._id);
+      res.writeHead(204, {'Content-Type': 'text/plain'});
+      res.write('the item known as  ' + req.url.query._id + 'has been deleted from storage');
+
       return;
     }
     res.writeHead(400, {'Content-Type': 'text/plain'});
